refactor(fields): remove unused suggestion logic and duplicate filtering
