+++ src/formatters/fileNameDisplayFormatter.ts
@@ -38,7 +38,7 @@ import {
 } from "../utils/generatedFilePath";
 import { getTemplateFile } from "../utils/templateFolderUtils";
 import { getValueVariableBaseName } from "../utils/valueSyntax";
-import { EnhancedFieldSuggestionFileFilter } from "../utils/EnhancedFieldSuggestionFileFilter";
+import { FieldSuggestionFileFilter } from "../utils/FieldSuggestionFileFilter";
 import { FILE_CUSTOM_PREFIX, FILE_PICK_PREFIX, type ParsedFileToken } from "../utils/fileSyntax";
 
 import type QuickAdd from "../main";
@@ -625,7 +625,7 @@ export class FileNameDisplayFormatter extends Formatter {
 	protected suggestForFile(parsed: ParsedFileToken): string {
 		// Preview: show a representative real file, else a placeholder. Never prompt.
 		const files = this.app
-			? EnhancedFieldSuggestionFileFilter.filterFiles(
+			? FieldSuggestionFileFilter.filterFiles(
 					this.app.vault.getMarkdownFiles(),
 					parsed.filter,
 					(file) => this.app!.metadataCache.getFileCache(file),
