+++ src/utils/FieldValueDeduplicator.test.ts
@@ -6,55 +6,38 @@ describe("FieldValueDeduplicator", () => {
 	describe("case-insensitive deduplication", () => {
 		it("should deduplicate basic case variations", () => {
 			const values = ["Active", "active", "ACTIVE", "Done", "DONE"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive',
-				preserveFirstOccurrence: true,
-				sortResult: false
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			expect(result.values).toEqual(["Active", "Done"]);
 			expect(result.duplicatesRemoved).toBe(3);
-			expect(result.strategy).toBe('case-insensitive');
 		});
 
 		it("should preserve first occurrence by default", () => {
 			const values = ["todo", "TODO", "Todo", "DONE", "done"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
-			// Should preserve "todo" (first) and "DONE" (first)
-			const resultSet = new Set(result.values);
-			expect(resultSet.has("todo") || resultSet.has("TODO") || resultSet.has("Todo")).toBe(true);
-			expect(resultSet.has("DONE") || resultSet.has("done")).toBe(true);
-			expect(result.values).toHaveLength(2);
+			expect(result.values).toEqual(["DONE", "todo"]);
 		});
 
 		it("should handle unicode normalization", () => {
 			const values = ["café", "cafe", "CAFÉ", "naïve", "naive"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
-			expect(result.values).toHaveLength(2);
+			expect(result.values).toEqual(["café", "naïve"]);
 			expect(result.duplicatesRemoved).toBe(3);
 		});
 
 		it("should handle empty strings and whitespace", () => {
 			const values = ["", " ", "test", "TEST", "   ", "test "];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			// Each whitespace variation should be considered different
-			expect(result.values.length).toBeGreaterThanOrEqual(4);
+			expect(new Set(result.values)).toEqual(new Set(["", " ", "test", "   ", "test "]));
 		});
 
 		it("should handle special characters and symbols", () => {
 			const values = ["@user", "@USER", "#tag", "#TAG", "user@domain", "USER@DOMAIN"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			expect(result.values).toHaveLength(3);
 			expect(result.duplicatesRemoved).toBe(3);
@@ -62,9 +45,7 @@ describe("FieldValueDeduplicator", () => {
 
 		it("should handle emoji and unicode characters", () => {
 			const values = ["🚀 Active", "🚀 ACTIVE", "✅ Done", "✅ done", "📝 Note"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			expect(result.values).toHaveLength(3);
 			expect(result.duplicatesRemoved).toBe(2);
@@ -72,9 +53,7 @@ describe("FieldValueDeduplicator", () => {
 
 		it("should preserve formatting differences (camelCase vs kebab-case)", () => {
 			const values = ["InProgress", "in-progress", "in_progress"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			// These should be considered different since they use different formatting conventions
 			expect(result.values).toHaveLength(3);
@@ -85,9 +64,7 @@ describe("FieldValueDeduplicator", () => {
 	describe("case-sensitive deduplication", () => {
 		it("should preserve case differences", () => {
 			const values = ["Active", "active", "ACTIVE", "Done"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-sensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values, true);
 
 			expect(result.values).toHaveLength(4);
 			expect(result.duplicatesRemoved).toBe(0);
@@ -95,34 +72,17 @@ describe("FieldValueDeduplicator", () => {
 
 		it("should remove exact duplicates only", () => {
 			const values = ["test", "test", "Test", "TEST"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-sensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values, true);
 
 			expect(result.values).toHaveLength(3);
 			expect(result.duplicatesRemoved).toBe(1);
 		});
 	});
 
-	describe("exact deduplication", () => {
-		it("should use Set-based deduplication", () => {
-			const values = ["same", "same", "different", "same"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'exact'
-			});
-
-			expect(result.values).toHaveLength(2);
-			expect(result.duplicatesRemoved).toBe(2);
-		});
-	});
-
 	describe("sorting behavior", () => {
 		it("should sort case-insensitively while preserving original case", () => {
 			const values = ["zebra", "Apple", "banana", "ZEBRA"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive',
-				sortResult: true
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			// Should be sorted: Apple, banana, zebra (case-insensitive order)
 			expect(result.values[0].toLowerCase()).toBe("apple");
@@ -130,27 +90,6 @@ describe("FieldValueDeduplicator", () => {
 			expect(result.values[2].toLowerCase()).toBe("zebra");
 		});
 
-		it("should prefer title case in sorting", () => {
-			const values = ["apple", "APPLE", "Apple"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive',
-				sortResult: true,
-				preserveFirstOccurrence: false
-			});
-
-			// Should preserve title case "Apple" over others
-			expect(result.values).toEqual(["Apple"]);
-		});
-
-		it("should disable sorting when requested", () => {
-			const values = ["zebra", "apple", "banana"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'exact',
-				sortResult: false
-			});
-
-			expect(result.values).toEqual(["zebra", "apple", "banana"]);
-		});
 	});
 
 	describePerf("performance with large datasets", () => {
@@ -162,9 +101,7 @@ describe("FieldValueDeduplicator", () => {
 			}
 
 			const startTime = performance.now();
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 			const endTime = performance.now();
 
 			expect(result.values).toHaveLength(100);
@@ -179,81 +116,11 @@ describe("FieldValueDeduplicator", () => {
 			}
 
 			const startTime = performance.now();
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 			const endTime = performance.now();
 
 			expect(result.values).toHaveLength(1000);
 			expect(endTime - startTime).toBeLessThan(500); // Should complete in < 500ms
-		});
-	});
-
-	describe("variation analysis", () => {
-		it("should analyze case variations correctly", () => {
-			const values = ["Active", "active", "ACTIVE", "Done", "done"];
-			const analysis = FieldValueDeduplicator.analyzeVariations(values);
-
-			expect(analysis.totalValues).toBe(5);
-			expect(analysis.uniqueValues).toBe(2);
-			expect(analysis.caseVariations.size).toBe(2);
-			
-			// Should identify the most common variation
-			expect(analysis.mostCommonCase).toBe("active");
-		});
-
-		it("should handle single variations", () => {
-			const values = ["unique1", "unique2", "unique3"];
-			const analysis = FieldValueDeduplicator.analyzeVariations(values);
-
-			expect(analysis.totalValues).toBe(3);
-			expect(analysis.uniqueValues).toBe(3);
-			expect(analysis.caseVariations.size).toBe(3);
-		});
-	});
-
-	describe("suggestions system", () => {
-		it("should suggest similar values", () => {
-			const existingValues = ["Active", "InProgress", "Done", "Cancelled"];
-			const suggestions = FieldValueDeduplicator.getSuggestions(
-				"activ", 
-				existingValues, 
-				0.7
-			);
-
-			expect(suggestions).toContain("Active");
-		});
-
-		it("should not suggest exact matches", () => {
-			const existingValues = ["Active", "InProgress", "Done"];
-			const suggestions = FieldValueDeduplicator.getSuggestions(
-				"Active", 
-				existingValues, 
-				0.8
-			);
-
-			expect(suggestions).not.toContain("Active");
-		});
-
-		it("should limit suggestions to 5 items", () => {
-			const existingValues = Array.from({length: 20}, (_, i) => `similar${i}`);
-			const suggestions = FieldValueDeduplicator.getSuggestions(
-				"simila", 
-				existingValues, 
-				0.5
-			);
-
-			expect(suggestions).toHaveLength(5);
-		});
-
-		it("should handle empty existing values", () => {
-			const suggestions = FieldValueDeduplicator.getSuggestions(
-				"test", 
-				[], 
-				0.8
-			);
-
-			expect(suggestions).toEqual([]);
 		});
 	});
 
@@ -274,9 +141,7 @@ describe("FieldValueDeduplicator", () => {
 
 		it("should handle arrays with only duplicates", () => {
 			const values = ["same", "SAME", "Same"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			expect(result.values).toHaveLength(1);
 			expect(result.duplicatesRemoved).toBe(2);
@@ -285,9 +150,7 @@ describe("FieldValueDeduplicator", () => {
 		it("should handle very long strings", () => {
 			const longString = "a".repeat(1000);
 			const values = [longString, longString.toUpperCase(), "short"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			expect(result.values).toHaveLength(2);
 			expect(result.duplicatesRemoved).toBe(1);
@@ -295,41 +158,11 @@ describe("FieldValueDeduplicator", () => {
 
 		it("should handle strings with newlines and special whitespace", () => {
 			const values = ["line1\nline2", "LINE1\nLINE2", "tab\there", "TAB\tHERE"];
-			const result = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive'
-			});
+			const result = FieldValueDeduplicator.deduplicate(values);
 
 			expect(result.values).toHaveLength(2);
 			expect(result.duplicatesRemoved).toBe(2);
 		});
 	});
 
-	describe("configuration options", () => {
-		it("should respect preserveFirstOccurrence setting", () => {
-			const values = ["First", "FIRST", "second"];
-			
-			const preserveFirst = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive',
-				preserveFirstOccurrence: true,
-				sortResult: false
-			});
-			
-			const preserveLast = FieldValueDeduplicator.deduplicate(values, {
-				strategy: 'case-insensitive',
-				preserveFirstOccurrence: false,
-				sortResult: false
-			});
-
-			expect(preserveFirst.values[0]).toBe("First");
-			expect(preserveLast.values[0]).toBe("FIRST");
-		});
-
-		it("should use default options when none provided", () => {
-			const values = ["test", "TEST"];
-			const result = FieldValueDeduplicator.deduplicate(values);
-
-			expect(result.strategy).toBe('case-insensitive');
-			expect(result.values).toHaveLength(1);
-		});
-	});
 });
